Flatten package commands
- Made all package commands base-line
