--- language/tools/move-cli/src/base/commands/mod.rs
@@ -1,9 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// Copyright (c) The Move Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod check;
-pub mod compile;
-
-pub use check::*;
-pub use compile::*;
