--- language/tools/move-cli/src/package/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) The Diem Core Contributors
-// Copyright (c) The Move Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod cli;
-pub mod prover;
