+++ language/tools/move-cli/tests/build_tests/circular_dependencies/args.txt
@@ -1,1 +1,1 @@
-package build -v
+build -v
