+++ language/tools/move-cli/tests/build_tests/dependency_chain/args.evm.txt
@@ -1,1 +1,1 @@
-package build -v --arch ethereum
+build -v --arch ethereum
