+++ language/tools/move-cli/tests/build_tests/dependency_chain/args.exp
@@ -1,4 +1,4 @@
-Command `package build -v`:
+Command `build -v`:
 INCLUDING DEPENDENCY Bar
 INCLUDING DEPENDENCY Foo
 BUILDING A
