+++ language/tools/move-cli/tests/build_tests/dependency_chain/args.txt
@@ -1,1 +1,1 @@
-package build -v
+build -v
