+++ language/tools/move-cli/tests/build_tests/dev_address/args.evm.txt
@@ -1,1 +1,1 @@
-package build -v -d --arch ethereum
+build -v -d --arch ethereum
