+++ language/tools/move-cli/tests/build_tests/dev_address/args.exp
@@ -1,2 +1,2 @@
-Command `package build -v -d`:
+Command `build -v -d`:
 BUILDING A
