+++ language/tools/move-cli/tests/build_tests/dev_address/args.txt
@@ -1,1 +1,1 @@
-package build -v -d
+build -v -d
