+++ language/tools/move-cli/tests/build_tests/empty_module_no_deps/args.exp
@@ -1,2 +1,2 @@
-Command `package build -v`:
+Command `build -v`:
 BUILDING A
