+++ language/tools/move-cli/tests/build_tests/empty_module_no_deps/args.txt
@@ -1,1 +1,1 @@
-package build -v
+build -v
