+++ language/tools/move-cli/tests/build_tests/include_exclude_stdlib/args.evm.txt
@@ -1,2 +1,2 @@
-package build -v --arch ethereum
--d -v package build --arch ethereum
+build -v --arch ethereum
+-d -v build --arch ethereum
