+++ language/tools/move-cli/tests/build_tests/only_deps/args.txt
@@ -1,3 +1,3 @@
-package build -v --only-deps
+build -v --only-deps
 > cat build/OnlyDeps/sources/A.move
 > ls build/OnlyDeps/sources/dependencies
