+++ language/tools/move-cli/tests/build_tests/rebuild_after_deleting_output_artifact/args.evm.txt
@@ -1,8 +1,8 @@
-package build -v --arch ethereum
-package build -v --arch ethereum
+build -v --arch ethereum
+build -v --arch ethereum
 > rm build/evm/Foo.yul
-package build -v --arch ethereum
+build -v --arch ethereum
 > rm build/evm/Foo.bin
-package build -v --arch ethereum
+build -v --arch ethereum
 > rm build/evm/Foo.abi.json
-package build -v --arch ethereum
+build -v --arch ethereum
