+++ language/tools/move-cli/tests/build_tests/rebuild_after_touching_manifest/args.evm.txt
@@ -1,3 +1,3 @@
-package build -v --arch ethereum
+build -v --arch ethereum
 > touch Move.toml
-package build -v --arch ethereum
+build -v --arch ethereum
