+++ language/tools/move-cli/tests/build_tests/rebuild_after_touching_source/args.evm.txt
@@ -1,3 +1,3 @@
-package build -v --arch ethereum
+build -v --arch ethereum
 > touch sources/Foo.move
-package build -v --arch ethereum
+build -v --arch ethereum
