+++ language/tools/move-cli/tests/build_tests/rebuild_no_modification/args.evm.txt
@@ -1,2 +1,2 @@
-package build -v --arch ethereum
-package build -v --arch ethereum
+build -v --arch ethereum
+build -v --arch ethereum
