+++ language/tools/move-cli/tests/build_tests/unbound_address/args.exp
@@ -1,4 +1,4 @@
-Command `package build -v`:
+Command `build -v`:
 Error: Unresolved addresses found: [
 Named address 'A' in package 'A'
 ]
