+++ language/tools/move-cli/tests/build_tests/unbound_dependency/args.evm.txt
@@ -1,1 +1,1 @@
-package build -v --arch ethereum
+build -v --arch ethereum
