+++ language/tools/move-cli/tests/build_tests/unbound_dependency/args.exp
@@ -1,4 +1,4 @@
-Command `package build -v`:
+Command `build -v`:
 Error: Unable to resolve packages for package 'A'
 
 Caused by:
