+++ language/tools/move-cli/tests/build_tests/unbound_dependency/args.txt
@@ -1,1 +1,1 @@
-package build -v
+build -v
