+++ language/tools/move-cli/tests/move_unit_tests/assign_dev_addr_for_dep/args.evm.txt
@@ -1,1 +1,1 @@
-package test --evm
+test --evm
