+++ language/tools/move-cli/tests/move_unit_tests/assign_dev_addr_for_dep/args.exp
@@ -1,4 +1,4 @@
-Command `package test`:
+Command `test`:
 INCLUDING DEPENDENCY Bar
 INCLUDING DEPENDENCY MoveStdlib
 BUILDING Foo
