+++ language/tools/move-cli/tests/move_unit_tests/assign_dev_addr_for_dep/args.stackless.exp
@@ -1,4 +1,4 @@
-Command `package test --stackless`:
+Command `test --stackless`:
 INCLUDING DEPENDENCY Bar
 INCLUDING DEPENDENCY MoveStdlib
 BUILDING Foo
