+++ language/tools/move-cli/tests/move_unit_tests/standalone_module_with_dev_addr_assignment/args.stackless.txt
@@ -1,1 +1,1 @@
-package test --stackless
+test --stackless
