+++ language/tools/move-cli/tests/move_unit_tests/standalone_module_with_regular_addr_assignment/args.exp
@@ -1,4 +1,4 @@
-Command `package test`:
+Command `test`:
 INCLUDING DEPENDENCY MoveStdlib
 BUILDING StandaloneModule
 Running Move unit tests
